requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
aiosqlite==0.19.0
aiofiles==23.2.1
Pillow==10.2.0
jinja2==3.1.3
pydantic==2.5.3
pydantic-settings==2.1.0
bcrypt==4.1.2
websockets==12.0
psutil==5.9.7
